+++ CHANGELOG.md
@@ -84,6 +84,10 @@ In addition, the "serve" CLI command has been improved to offer a self-hosted ve
 - `gr`: Corrected key for the reduced island tax rate.
 - `bill`: Payment Line tax always calculated.
 
+### Added
+
+- `ad`: Andorra regime
+
 ## [v0.308.0] - 2026-02-17
 
 ### Added